requirements.txt
python-dotenv~=1.0.0
openai~=1.60.0
anthropic~=0.45.0
transformers~=4.44.2